refactor(permissions): Move permissions to constants
