--- src/main/kotlin/uno/rebellious/twitchbot/command/Permission.kt
@@ -1,3 +1,0 @@
-package uno.rebellious.twitchbot.command
-
-data class Permission(val isOwnerOnly: Boolean, val isModOnly: Boolean, val isSubOnly: Boolean)
